moved header to separate component
