+++ app/root.tsx
@@ -61,7 +61,6 @@ export function Layout({ children }: { children: React.ReactNode }) {
   const defaultColorScheme = rootLoaderData?.requestInfo?.colorScheme;
   const username = rootLoaderData?.requestInfo?.user?.username;
   const [colorScheme, setColorScheme] = useState(defaultColorScheme);
-  const submit = useSubmit();
 
   return (
     <html lang="en" className={colorScheme}>
@@ -72,55 +71,8 @@ export function Layout({ children }: { children: React.ReactNode }) {
         <Links />
       </head>
       <body className="h-screen w-full overscroll-none">
-        <Header />
-        {/* <div className="flex flex-row items-center space-x-4">
-          <Link to="/">
-            <img
-              src="/logo.png"
-              alt="Logo"
-              className="h-16 w-16 inline-block"
-            />
-          </Link>
-          <H1 className="grow">Tito and Dom's M+ Adventure</H1>
-          <div className="flex flex-row space-x-2">
-            {username ? (
-              <>
-                <span>Hi {username}</span>
-                <Form action="/auth/signout" method="post">
-                  <Button type="submit">Sign Out</Button>
-                </Form>
-              </>
-            ) : (
-              <>
-                <Button asChild>
-                  <Link to="/auth/signin">Sign In</Link>
-                </Button>
-                <Button asChild>
-                  <Link to="/auth/signup">Sign Up</Link>
-                </Button>
-              </>
-            )}
-          </div>
-          <div className="flex flex-row items-center space-x-2">
-            <Label>Dark Mode</Label>
-            <Switch
-              checked={colorScheme === "dark"}
-              onCheckedChange={(v) => {
-                const newValue = v ? "dark" : "light";
-                setColorScheme(newValue);
-                submit(
-                  { colorScheme: newValue },
-                  {
-                    method: "post",
-                    navigate: false,
-                    action: "/",
-                    encType: "application/json",
-                  }
-                );
-              }}
-            />
-          </div>
-        </div> */}
+        <Header colorScheme={colorScheme} setColorScheme={setColorScheme} username={username} />
+
         {children}
         <ScrollRestoration />
         <Scripts />
